requirements.txt
flet==0.22.*
requests
mysql-connector-python
pyinstaller
pillow